pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ghexplain"
version = "0.0.3"
authors = [
    { name = "Stefano Fancello", email = "stefano.fancello@nethesis.it" },
]
description = "AI-powered explanations for GitHub issues using LangChain"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0"
]

[project.urls]
"Homepage" = "https://github.com/Stell0/ghexplain"
"Bug Tracker" = "https://github.com/Stell0/ghexplain/issues"

[project.scripts]
ghexplain = "ghexplain.cli:main"

[tool.setuptools]
packages = ["ghexplain"]
package-dir = {"" = "src"}
